requirements.txt
streamlit
PyPDF2
langchain
langchain-community
faiss-cpu
sentence-transformers
huggingface-hub
python-dotenv
requests
